[clear]frontのリファクタリング
page.tsxに固まりすぎていたのでコンポーネント化を実施（途中）
