+++ front/lib/atoms/currentCityState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const currentCityStat = atom({
+  key: 'currentCityStat',
+  default: '',
+});
