+++ front/lib/atoms/dayWeatherState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const dayWeatherState = atom({
+  key: 'dayWeatherState',
+  default: [],
+});
