+++ front/lib/atoms/disabledState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const disabledState = atom({
+  key: 'disabledState',
+  default: true,
+});
