+++ front/lib/atoms/errorMessageState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const errorMessageState = atom({
+  key: 'errorMessageState',
+  default: '',
+});
