+++ front/lib/atoms/getErrorState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const getErrorState = atom({
+  key: 'getError',
+  default: false,
+});
