+++ front/lib/atoms/inputCityState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const inputCityState = atom({
+  key: 'inputCityState',
+  default: '',
+});
