+++ front/lib/atoms/isLoadingState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const isLoadingState = atom({
+  key: 'isLoading',
+  default: false,
+});
