+++ front/lib/atoms/timeWeatherState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const timeWeatherState = atom({
+  key: 'timeWeatherState',
+  default: [],
+});
