+++ front/lib/atoms/weatherDataState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const weatherDataState: any = atom({
+  key: 'weatherDataState',
+  default: [],
+});
