+++ front/package.json
@@ -21,6 +21,7 @@
     "postcss": "8.4.29",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "recoil": "^0.7.7",
     "swr": "^2.2.2",
     "tailwindcss": "3.3.3",
     "typescript": "5.2.2"
